fix: exllamav2 fix attempt

+++ kernels/quantization/exl2/q_gemm_exl2.cu
@@ -119,7 +119,7 @@ void gemm_half_q_half_cuda(cublasHandle_t cublas_handle, const half* a,
 }  // namespace exl2
 }  // namespace aphrodite
 
-torch::Tensor exl2_gemm(torch::Tensor a, uintptr_t b) {
+torch::Tensor exl2_gemm(torch::Tensor a, int64_t b) {
   const at::cuda::OptionalCUDAGuard device_guard(device_of(a));
   aphrodite::exl2::QMatrix* qm = reinterpret_cast<aphrodite::exl2::QMatrix*>(b);
 
@@ -140,7 +140,7 @@ torch::Tensor exl2_gemm(torch::Tensor a, uintptr_t b) {
   return c;
 }
 
-uintptr_t make_q_matrix(torch::Tensor q_weight, torch::Tensor q_perm,
+int64_t make_q_matrix(torch::Tensor q_weight, torch::Tensor q_perm,
                         torch::Tensor q_invperm, torch::Tensor q_scale,
                         torch::Tensor q_scale_max, torch::Tensor q_groups,
                         torch::Tensor q_group_map) {
